socialnetworkapp/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.2
colorama==0.4.6
cryptography==44.0.1
dataclasses-json==0.6.7
defusedxml==0.7.1
diff-match-patch==20241021
distro==1.9.0
Django==5.1.6
django-ckeditor==6.7.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-filter==25.1
django-import-export==4.3.8
django-js-asset==3.0.1
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docx2txt==0.9
drf-yasg==1.21.9
faiss-cpu==1.12.0
filelock==3.19.1
firebase-admin==6.8.0
frozenlist==1.7.0
fsspec==2025.7.0
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jwcrypto==1.5.6
kombu==5.5.3
langchain==0.2.17
langchain-community==0.2.17
langchain-core==0.2.43
langchain-huggingface==0.0.3
langchain-openai==0.1.7
langchain-text-splitters==0.2.2
langsmith==0.1.112
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
numpy==1.26.4
oauthlib==3.2.2
openai==1.100.2
orjson==3.11.2
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.3
pypdf==6.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
qrcode==8.1
regex==2025.7.34
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.35.0
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
sympy==1.14.0
tablib==3.8.0
tenacity==8.3.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.4
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
zstandard==0.24.0
